santa_demo/requirements.txt
websockets
sounddevice
numpy
requests
dotenv
tiktoken
openwakeword
deepgram-sdk==2.12.0
pyaudio
pydub
agents
openai_agents
speexdsp